chore: refactor action functions
